+++ lib/gp2map.gd
@@ -137,4 +137,10 @@ DeclareOperation( "IsomorphismPreCat1Groups",
 
 #############################################################################
 ##
+#A  MorphismOfPullback( <xmod> )
+##
+DeclareAttribute( "MorphismOfPullback", IsXMod );
+
+#############################################################################
+##
 #E  gp2map.gd . . . . . . . . . . . . . . . . . . . . . . . . . . . ends here