+++ lib/map2arg.gd
@@ -2,18 +2,20 @@
 ##
 #W  map2arg.gd                   XMOD Package                    Chris Wensley
 ##
-#Y  Copyright (C) 2001-2016, Chris Wensley et al,  
+#Y  Copyright (C) 2001-2018, Chris Wensley et al,  
 #Y  School of Computer Science, Bangor University, U.K. 
 
 #############################################################################
 ##                                         
-#F  Mapping2ArgumentsByFunction( <S>, <R>, <fun> ) . create map from function
-#F  IsMapping2ArgumentsByFunction
-##
-##  `Mapping2ArgumentsByFunction' creates a  `IsNonSPGeneralMapping' ??
+#O  Mapping2ArgumentsByFunction( <S>, <R>, <fun> ) . create map from function
+#P  IsMapping2ArgumentsByFunction( <map> ) 
+#O  ImageElmMapping2ArgumentsByFunction( <map>, <pair> ) 
 ##              
-DeclareGlobalFunction( "Mapping2ArgumentsByFunction" );
-DeclareProperty( "IsMapping2ArgumentsByFunction", IsMappingByFunctionRep );
+DeclareOperation( "Mapping2ArgumentsByFunction", 
+    [ IsList, IsDomain, IsFunction ] );
+DeclareProperty( "IsMapping2ArgumentsByFunction", IsGeneralMapping );
+DeclareOperation( "ImageElmMapping2ArgumentsByFunction", 
+    [ IsMapping2ArgumentsByFunction, IsList ] );
 
 #############################################################################
 ##